+++ apps/native-component-list/src/screens/ModulesCore/ExpoModulesScreen.tsx
@@ -8,6 +8,10 @@ const customJsonReplacer = (_: string, value: any) => {
   return typeof value === 'function' ? value.toString().replace(/\s+/g, ' ') : value;
 };
 
+function CheckIfPresent({ name }: { name: string }) {
+  return <MonoText>{`'${name}' in globalThis.expo => ${name in globalThis.expo!}`}</MonoText>;
+}
+
 export default function ExpoModulesScreen() {
   const modules = { ...globalThis.expo?.modules };
   const moduleNames = Object.keys(modules);
@@ -15,7 +19,12 @@ export default function ExpoModulesScreen() {
   return (
     <ScrollView style={styles.scrollView}>
       <HeadingText>Host object is installed</HeadingText>
-      <MonoText>{`'modules' in globalThis.expo => ${'modules' in globalThis.expo!}`}</MonoText>
+      <CheckIfPresent name="modules" />
+
+      <HeadingText>Exported classes</HeadingText>
+      {['EventEmitter', 'SharedObject', 'SharedRef', 'NativeModule'].map((className) => (
+        <CheckIfPresent key={className} name={className} />
+      ))}
 
       <HeadingText>Available Expo modules</HeadingText>
       <MonoText>