+++ packages/expo-location/src/Location.ts
@@ -1,3 +1,4 @@
+import { isRunningInExpoGo } from 'expo';
 import { PermissionResponse, createPermissionHook, Platform } from 'expo-modules-core';
 
 import ExpoLocation from './ExpoLocation';
@@ -18,6 +19,9 @@ import {
   LocationTaskOptions,
 } from './Location.types';
 import { LocationSubscriber, HeadingSubscriber } from './LocationSubscribers';
+
+// Flag for warning about background services not being available in Expo Go
+let warnAboutExpoGoDisplayed = false;
 
 // @needsAudit
 /**
@@ -313,9 +317,20 @@ export async function hasServicesEnabledAsync(): Promise<boolean> {
 
 // --- Background location updates
 
-function _validateTaskName(taskName: string) {
+function _validate(taskName: string) {
   if (!taskName || typeof taskName !== 'string') {
     throw new Error(`\`taskName\` must be a non-empty string. Got ${taskName} instead.`);
+  }
+  if (isRunningInExpoGo()) {
+    if (!warnAboutExpoGoDisplayed) {
+      const message =
+        'Background location is limited in Expo Go:\n' +
+        'On Android, it is not available at all.\n' +
+        'On iOS, it works when running in the Simulator.\n' +
+        'Please use a development build to avoid limitations. Learn more: https://expo.fyi/dev-client.';
+      console.warn(message);
+      warnAboutExpoGoDisplayed = true;
+    }
   }
 }
 
@@ -356,7 +371,7 @@ export async function startLocationUpdatesAsync(
   taskName: string,
   options: LocationTaskOptions = { accuracy: LocationAccuracy.Balanced }
 ): Promise<void> {
-  _validateTaskName(taskName);
+  _validate(taskName);
   await ExpoLocation.startLocationUpdatesAsync(taskName, options);
 }
 
@@ -367,7 +382,7 @@ export async function startLocationUpdatesAsync(
  * @return A promise resolving as soon as the task is unregistered.
  */
 export async function stopLocationUpdatesAsync(taskName: string): Promise<void> {
-  _validateTaskName(taskName);
+  _validate(taskName);
   await ExpoLocation.stopLocationUpdatesAsync(taskName);
 }
 
@@ -378,7 +393,7 @@ export async function stopLocationUpdatesAsync(taskName: string): Promise<void> 
  * started or not.
  */
 export async function hasStartedLocationUpdatesAsync(taskName: string): Promise<boolean> {
-  _validateTaskName(taskName);
+  _validate(taskName);
   return ExpoLocation.hasStartedLocationUpdatesAsync(taskName);
 }
 
@@ -446,7 +461,7 @@ export async function startGeofencingAsync(
   taskName: string,
   regions: LocationRegion[] = []
 ): Promise<void> {
-  _validateTaskName(taskName);
+  _validate(taskName);
   _validateRegions(regions);
   await ExpoLocation.startGeofencingAsync(taskName, { regions });
 }
@@ -459,7 +474,7 @@ export async function startGeofencingAsync(
  * @return A promise resolving as soon as the task is unregistered.
  */
 export async function stopGeofencingAsync(taskName: string): Promise<void> {
-  _validateTaskName(taskName);
+  _validate(taskName);
   await ExpoLocation.stopGeofencingAsync(taskName);
 }
 
@@ -470,6 +485,6 @@ export async function stopGeofencingAsync(taskName: string): Promise<void> {
  * started or not.
  */
 export async function hasStartedGeofencingAsync(taskName: string): Promise<boolean> {
-  _validateTaskName(taskName);
+  _validate(taskName);
   return ExpoLocation.hasStartedGeofencingAsync(taskName);
 }