+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/functions/SyncFunctionComponent.kt
@@ -9,16 +9,17 @@ import expo.modules.kotlin.jni.JNIFunctionBody
 import expo.modules.kotlin.jni.decorators.JSDecoratorsBridgingObject
 import expo.modules.kotlin.types.AnyType
 import expo.modules.kotlin.types.JSTypeConverter
+import expo.modules.kotlin.types.ReturnType
 
 class SyncFunctionComponent(
   name: String,
-  desiredArgsTypes: Array<AnyType>,
+  argTypes: Array<AnyType>,
+  private val returnType: ReturnType,
   private val body: (args: Array<out Any?>) -> Any?
-) : AnyFunction(name, desiredArgsTypes) {
+) : AnyFunction(name, argTypes) {
   private var shouldUseExperimentalConverter = false
 
-  @Suppress("FunctionName")
-  fun UseExperimentalConverter(shouldUse: Boolean = true) = apply {
+  fun useExperimentalConverter(shouldUse: Boolean = true) = apply {
     shouldUseExperimentalConverter = shouldUse
   }
 
@@ -37,7 +38,11 @@ class SyncFunctionComponent(
         FunctionCallException(name, moduleName, it)
       }) {
         val result = call(args, appContext)
-        return@exceptionDecorator JSTypeConverter.convertToJSValue(result, useExperimentalConverter = shouldUseExperimentalConverter)
+        if (shouldUseExperimentalConverter) {
+          return@exceptionDecorator returnType.convertToJS(result)
+        } else {
+          return@exceptionDecorator JSTypeConverter.convertToJSValue(result)
+        }
       }
     }
   }
@@ -46,6 +51,7 @@ class SyncFunctionComponent(
     jsObject.registerSyncFunction(
       name,
       takesOwner,
+      isEnumerable,
       getCppRequiredTypes().toTypedArray(),
       getJNIFunctionBody(moduleName, appContext)
     )