+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/notifications/interfaces/NotificationTrigger.java
@@ -1,10 +1,8 @@
 package expo.modules.notifications.notifications.interfaces;
 
-import android.os.Build;
 import android.os.Parcelable;
 
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 
 /**
  * An interface specifying source of the notification, to be implemented
@@ -12,7 +10,6 @@ import androidx.annotation.RequiresApi;
  */
 public interface NotificationTrigger extends Parcelable {
   @Nullable
-  @RequiresApi(api = Build.VERSION_CODES.O)
   default String getNotificationChannel() {
     return null;
   }